Fix Cohere models by using new `langchain-cohere` partner package (#848)
* fix Cohere models by depending on langchain_cohere

* remove Cohere import

* update Cohere provider dependencies in docs
