+++ packages/jupyter-ai-magics/jupyter_ai_magics/__init__.py
@@ -4,7 +4,6 @@ from ._version import __version__
 from .embedding_providers import (
     BaseEmbeddingsProvider,
     BedrockEmbeddingsProvider,
-    CohereEmbeddingsProvider,
     GPT4AllEmbeddingsProvider,
     HfHubEmbeddingsProvider,
     QianfanEmbeddingsEndpointProvider,
@@ -22,7 +21,6 @@ from .providers import (
     BaseProvider,
     BedrockChatProvider,
     BedrockProvider,
-    CohereProvider,
     GPT4AllProvider,
     HfHubProvider,
     QianfanProvider,
