+++ packages/jupyter-ai-magics/jupyter_ai_magics/providers.py
@@ -36,7 +36,6 @@ from langchain_community.chat_models import BedrockChat, QianfanChatEndpoint
 from langchain_community.llms import (
     AI21,
     Bedrock,
-    Cohere,
     GPT4All,
     HuggingFaceEndpoint,
     SagemakerEndpoint,
@@ -544,26 +543,6 @@ class AI21Provider(BaseProvider, AI21):
         if isinstance(e, ValueError):
             return "status code 401" in str(e)
         return False
-
-
-class CohereProvider(BaseProvider, Cohere):
-    id = "cohere"
-    name = "Cohere"
-    # Source: https://docs.cohere.com/reference/generate; https://docs.cohere.com/docs/models
-    models = [
-        "command",
-        "command-nightly",
-        "command-light",
-        "command-light-nightly",
-        "command-r-plus",
-        "command-r",
-    ]
-    model_id_key = "model"
-    pypi_package_deps = ["cohere"]
-    auth_strategy = EnvAuthStrategy(name="COHERE_API_KEY")
-
-    async def _acall(self, *args, **kwargs) -> Coroutine[Any, Any, str]:
-        return await self._call_in_executor(*args, **kwargs)
 
 
 class GPT4AllProvider(BaseProvider, GPT4All):
